Refactoring the testing directory structure
