+++ tests/Storage/SimpleCacheStorageTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Yii\RateLimiter\Tests;
+namespace Yiisoft\Yii\RateLimiter\Tests\Storage;
 
 use InvalidArgumentException;
 use Yiisoft\Cache\ArrayCache;
